Set view permissions to explicit roles from the parent table
